monolith/src/catalog/api/v1/serializers.py
from rest_framework import serializers

from catalog.models import CatalogEntry


class CatalogSerializer(serializers.ModelSerializer):

    class Meta:
        model = CatalogEntry
        fields = '__all__'