monolith/requirements.txt
django
django_rest_framework
django-extensions
celery
gunicorn
psycopg2
python-dateutil
pykafka